refactor: re-organize the API
